Fix handling of short-lived processes
When building the process tree, it lists all processes then tries to get
info about each in turn from the disk, this isn't instant and sometimes
processes exit in between, so we have to handle the case of a process
not existing anymore when we look it up. We do that by ignoring that
process which exited.